+++ CHANGELOG.md
@@ -1,28 +1,52 @@
 1.2.17 (unreleased)
+
+- Feature #584: upload model upper limit 2G \nFeature #650: SDK get influxdb credential \nFeature #632: AFS upload model Naming rule
+
 -------------------
 
-- Add Feature #584: AFS upload model upper limit 200MB\nAdd Feature #632: AFS upload model naming rule and unit test
-
-- Nothing changed yet.
+
+
+- Add Feature #584: AFS upload model upper limit 200MB
+
+- Add Feature #632: AFS upload model naming rule and unit test
 
 
 1.2.16 (2018-08-22)
 
+
+
 -------------------
 
+
+
 - Add Feature #599: AFS SDK get numerical, target, select_feature
+
 - Add Feature #574: Add SDK can set and get parameter by list type.
+
 - Add Feature #595: Integrate SDK environment variable and flow handling
+
 - Add Feature: Add SDK services modules, get credential from the subscribed databases or services.
+
+
+
 
 
 1.2.15 (unreleased)
 
+
+
 -------------------
+
+
+
 
 
 1.2.14 (2018-08-10)
 
+
+
 -------------------
+
+
 
 - Add Feature #542 - add get_join_table documents